+++ lib/searchkick/record_indexer.rb
@@ -20,7 +20,14 @@ module Searchkick
           raise Searchkick::Error, "Partial reindex not supported with queue option"
         end
 
-        index.reindex_queue.push(record.id.to_s)
+        # always pass routing in case record is deleted
+        # before the queue job runs
+        if record.respond_to?(:search_routing)
+          routing = record.search_routing
+        end
+
+        value = [record.id.to_s, routing].join("|")
+        index.reindex_queue.push(value)
       when :async
         unless defined?(ActiveJob)
           raise Searchkick::Error, "Active Job not found"