+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.1.2 [unreleased]
 
 - Fixed deletes with routing and `async` callbacks
+- Fixed deletes with routing and `queue` callbacks
 
 ## 3.1.1